refactor: Finder * fixed map expander crash at Finder page
